requirements.txt
google-genai
python-dotenv
pypdf
langchain-text-splitters
faiss-cpu